refactor: moved data folder
